Refactor how routing is managed
Split the single route /results into /test-results, /result and
/measurement.

This makes is more straightforward to manage the various views and makes
it possible to nest the components specific to each page inside of it's
own directory tree.
