+++ renderer/components/UploadSpeed.js
