+++ renderer/components/VideoQuality.js
