+++ renderer/components/formatBitrate.js
