+++ renderer/components/formatSpeed.js
